files source code

+++ src/main/java/com/becoder/repository/HotelRepository.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface HotelRepository extends JpaRepository<Hotels, Long> {
-
+public interface HotelRepository extends JpaRepository<Hotels, Integer> {
 }
