+++ src/main/java/com/becoder/repository/PlacesRepository.java
@@ -5,6 +5,8 @@ import com.becoder.model.Places;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface PlacesRepository extends JpaRepository<Places,Long> {
+public interface PlacesRepository extends JpaRepository<Places,Integer> {
 }
